requirements.txt
python-docx
pdfplumber
pandas
openpyxl
matplotlib
pytest
